frontend-web: Fix some TS4023: cannot be named errors
As described here: https://github.com/Microsoft/TypeScript/issues/5711
typescript has a nasty feature where implicitly using an imported type
in the definition of an export causes a TS4023 error.

One suggested solution is to make the type of the export expicit.
(This also helps make the code a bit more comprehensible.)

This change uses that solution to fix a bunch of compile errors.