frontend-web: Add key field to various elements
This fixes react "array elements must have keys" errors.
Mostly this is required when the list of elements is dynamically
generated - react uses the key to work out which element has
been added/removed.

(In some cases, I suspect the error is actually caused by a bug
in react as the elements are not dynamically generated.)