frontend: some package version churn.
Using a proper import for redux-thunk exposed the fact that this
version of the library does not support our version of redux.

And the redux update is (probably) blocked until we can replace
react-router-redux, which is no longer supported.

Unfortunately, I have no idea what react-router-redux is actually
doing, which makes it difficult to replace:-(

So in the meantime, downgrade redux-thunk to a supported version.
Clean up some other dependencies at the same time.